Add Missing Water Entry Points
There are a few things that still need to be considered. In Reach some of these entry points do have a per vertex lighting flag, which has not been implemented yet. The static_sh entry point while it is included in the render method definition, there are no HLSL functions or macros which reference it. It does have a shader in the ODSTEK glvs, however it is empty.