Add Initial Single Pass Entry Support
There is a lot that still needs to be worked on here. The `SINGLE_PASS_LIGHTING` macro still needs to be implemented correctly, and since foliage uses its own implementation, that also needs to be transferred over from the reach source. Of minor note is the active_camo entry point inside the halogram shader, which uses the reach active_camo code. That will also need to be pulled from the reach source